requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
httpx==0.28.1
playwright==1.49.1
playwright-stealth==1.0.6
pydantic==2.10.6
python-dotenv==1.0.1
undetected-chromedriver==3.5.5
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
